tests: system util tests refactoring to remove impossible scenarios
The tests for the system allocs reconciling code path (`diffSystemAllocs`)
include many impossible test environments, such as passing allocs for the wrong
node into the function. This makes the test assertions nonsensible for use in
walking yourself through the correct behavior. This commit breaks up a couple of
tests, expands test coverage, and makes test assertions more clear.

This commit also clarifies the purpose of `materializeSystemTaskGroups`, which
misleadingly reads the task group count (which for system jobs is always 0).