scheduler: system scheduler should reconcile overlapping live allocs
Under conditions that are not yet understood, it's possible for there to be
overlapping live system job allocations on the same node. The scheduler does not
reconcile this condition by stopping an extra allocation, but the extra also
does not show up in the plan.

When allocations need exclusive resources such as reserved ports, this results
in the plan applier rejecting the plan. As far as the scheduler knows the plan
is good, and because this is for a system job, the scheduler will repeatedly
submit this bad plan. This results in the dreaded "plan for node rejected"
loop. Although we should definitely figure out why we get into this bad state to
begin with, it's the job of the scheduler to properly reconcile the cluster
state even in the face of errors.

Add a new `ensureSingleSystemAlloc` function that enforces the invariant that
there can be only a single desired-running allocation for a given system job on
a given node.